django_metabase/requirements.txt
PyJWT==2.9.0
python-dotenv==1.0.1
django==5.2.5